+++ riscv-macros/src/lib.rs
@@ -4,6 +4,9 @@ use syn::{parse_macro_input, DeriveInput};
 
 mod riscv;
 
+#[cfg(feature = "riscv-rt")]
+mod riscv_rt;
+
 /// Attribute-like macro that implements the traits of the `riscv-types` crate for a given enum.
 ///
 /// As these traits are unsafe, the macro must be called with the `unsafe` keyword followed by the trait name.
@@ -58,4 +61,49 @@ pub fn pac_enum(attr: TokenStream, item: TokenStream) -> TokenStream {
         #(#trait_impl)*
     }
     .into()
+}
+
+/// Temporary patch macro to deal with LLVM bug.
+///
+/// # Note
+///
+/// This macro is intended to be used internally by the `riscv-rt` crate. Do not use it directly in your code.
+#[cfg(feature = "riscv-rt")]
+#[proc_macro]
+pub fn rvrt_llvm_arch_patch(_input: TokenStream) -> TokenStream {
+    let q = if let Ok(arch) = std::env::var("RISCV_RT_LLVM_ARCH_PATCH") {
+        let patch = format!(".attribute arch,\"{arch}\"");
+        quote! { core::arch::global_asm!{#patch} }
+    } else {
+        quote!(compile_error!("RISCV_RT_LLVM_ARCH_PATCH is not set"))
+    };
+    q.into()
+}
+
+/// Generates assembly code required for the default handling of traps.
+///
+/// The main routine generated is `_default_start_trap`. If no `_start_trap` function
+/// is defined, the linker will use this function as the default trap entry point.
+///
+/// If the `pre-default-start-trap` feature is enabled, the generated code will also
+/// include a call to a user-defined function `_pre_default_start_trap` at the beginning
+/// of the `_default_start_trap` routine.
+///
+/// If the `rt-v-trap` feature is enabled, the macro will also include the assembly code
+/// for the `_start_DefaultInterrupt_trap` and `_continue_interrupt_trap` routines, which
+/// are required for handling core interrupts in vectored trap mode.
+///
+/// # Note
+///
+/// This macro is intended to be used internally by the `riscv-rt` crate. Do not use it directly in your code.
+#[cfg(feature = "riscv-rt")]
+#[proc_macro]
+pub fn rvrt_default_start_trap(_input: TokenStream) -> TokenStream {
+    match riscv_rt::asm::RiscvArch::try_from_env() {
+        Some(arch) => arch.default_start_trap().into(),
+        None => quote! {
+            compile_error!("RISCV_RT_BASE_ISA environment variable is not set or is invalid");
+        }
+        .into(),
+    }
 }